Fix gemma3 export from a vlm checkpoint (#90)
@thewh1teagle this PR allows to correctly export with
text-generation(-with-past) task even when loading from a vlm
checkpoint.
it in the last PR there was only a causal lm checkpoint in the
onnxruntime tests, and a vlm checkpoint in the exporters tests (which I
wasn't aware of its failure)

@echarlaix apparently with gemma3, there's no way to make the model
output the past_key_values (even when all configs and subconfigs have
their use_cache attribute set to True) because of [this
line](https://github.com/huggingface/transformers/blob/v4.55.4/src/transformers/models/gemma3/modeling_gemma3.py#L951)
I have made changes so that we can pass use_cache as an argument
instead, this seems to be more reliable and removes the need for
patching a couple models where a specific use_cache had to be set
(seq2seq models and trocr).